# openapi.yaml
openapi: 3.0.3
info:
  title: MVTMS REST API
  description: |
    OpenAPI specification for the MVTMS REST API v1. Authentication is via
    a signed Bearer token (`Authorization: Bearer <token>`) or an existing
    session cookie. The token endpoints are public; all other `/api/v1/*`
    routes require a valid principal plus Casbin RBAC permissions.
  version: "1.0.0"
  contact:
    name: MVTMS Support
    url: https://avandab.com
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/token:
    post:
      tags: [Auth]
      summary: Issue API token
      description: Authenticate with email and password to receive a signed Bearer token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Signed token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Public self-registration. Only the `viewer` role can be requested;
        privileged roles must be assigned by an administrator.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered and token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed API token returned by `POST /api/v1/auth/token`.
...
  schemas:
...
    TokenRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user_id:
          type: string
        role:
          type: string
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          format: password
        role:
          type: string
          description: Only `viewer` is allowed for public registration.
          default: viewer
        vehicle_number:
          type: string
    RegisterResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or invalid credentials.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient RBAC permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...